// src/lib/api/generated/model/publicForm.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { Status37cEnum } from './status37cEnum';
import type { PublicFormQuestion } from './publicFormQuestion';

export interface PublicForm {
  readonly id: string;
  readonly title: string;
  readonly description: string;
  readonly status: Status37cEnum;
  readonly questions: readonly PublicFormQuestion[];
}