// src/lib/api/generated/model/publicFormQuestion.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionTypeEnum } from './questionTypeEnum';
import type { PublicFormQuestionOption } from './publicFormQuestionOption';

export interface PublicFormQuestion {
  readonly id: string;
  readonly label: string;
  readonly question_type: QuestionTypeEnum;
  readonly is_required: boolean;
  readonly order: number;
  readonly options: readonly PublicFormQuestionOption[];
}