// src/lib/api/generated/model/publicSurveySubmit.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { PublicSurveyAnswerInput } from './publicSurveyAnswerInput';

export interface PublicSurveySubmit {
  answers: PublicSurveyAnswerInput[];
}