// src/lib/api/generated/model/questionTypeEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `TEXT` - Text
* `SINGLE_CHOICE` - Single choice
* `MULTIPLE_CHOICE` - Multiple choice
* `SCALE_1_5` - Scale 1-5
* `SCALE_0_10` - Scale 0-10
 */
export type QuestionTypeEnum = typeof QuestionTypeEnum[keyof typeof QuestionTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const QuestionTypeEnum = {
  TEXT: 'TEXT',
  SINGLE_CHOICE: 'SINGLE_CHOICE',
  MULTIPLE_CHOICE: 'MULTIPLE_CHOICE',
  SCALE_1_5: 'SCALE_1_5',
  SCALE_0_10: 'SCALE_0_10',
} as const;